Refactor plugin provider into parent and child classes
